requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
requests==2.31.0
psycopg2-binary==2.9.9
cryptography==42.0.5
twilio==9.0.4
python-multipart==0.0.9
pydantic==2.7.1
